+++ db/playsms.sql
@@ -1229,7 +1229,7 @@ CREATE TABLE `playsms_tblSMSOutgoing_queue` (
   `datetime_scheduled` varchar(20) NOT NULL DEFAULT '000-00-00 00:00:00',
   `datetime_update` varchar(20) NOT NULL DEFAULT '000-00-00 00:00:00',
   `flag` tinyint(4) NOT NULL,
-  `sms_count` tinyint(4) NOT NULL,
+  `sms_count` int(11) NOT NULL,
   `uid` int(11) NOT NULL,
   `gpid` int(11) NOT NULL,
   `sender_id` varchar(100) NOT NULL,
